backend/requirements.txt
anaconda==0.0.1.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.5
bottle==0.12.18
bs4==0.0.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
conda==4.3.16
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
graphviz==0.14
gspread==3.6.0
httplib2==0.18.1
idna==2.10
install==1.3.4
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.3
jupyterlab-server==1.1.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lfr==0.1.2
MarkupSafe==1.1.1
marshmallow==3.6.0
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.5
notebook==6.0.3
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parchmint==0.1.2
parso==0.7.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
platformio==4.3.4
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pyelftools==0.26
Pygments==2.6.1
pygraphviz==1.6
pylint==2.5.2
pymint==0.1.2
pyparsing==2.4.7
pyrsistent==0.16.0
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
semantic-version==2.8.5
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stevedore==2.0.1
svgwrite==1.4
tabula==1.0.5
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.0.25
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarn==0.0.8
